Relocate IDocumentTextExtractor and its 3 implementations to Shared.Rag
