+++ backend/src/Anela.Heblo.Application/Features/KnowledgeBase/Services/DocumentIndexingService.cs
@@ -1,3 +1,4 @@
+using Anela.Heblo.Application.Shared.Rag;
 using Anela.Heblo.Domain.Features.KnowledgeBase;
 using Anela.Heblo.Domain.Shared.Rag;
