+++ backend/src/Anela.Heblo.Application/Shared/Rag/DocumentExtractors/WordDocumentExtractor.cs
@@ -2,7 +2,7 @@ using DocumentFormat.OpenXml.Packaging;
 using DocumentFormat.OpenXml.Wordprocessing;
 using Microsoft.Extensions.Logging;
 
-namespace Anela.Heblo.Application.Features.KnowledgeBase.Services.DocumentExtractors;
+namespace Anela.Heblo.Application.Shared.Rag.DocumentExtractors;
 
 public class WordDocumentExtractor : IDocumentTextExtractor
 {
